requirements.txt
together==0.2.5
sentence-transformers==2.5.1
faiss-cpu==1.7.4
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
flask>=3.0.0
flask-session>=0.5.0
python-dotenv>=1.0.0